% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMix_se.R
\name{MRMix_se}
\alias{MRMix_se}
\title{Standard error of the MRMix estimator}
\usage{
MRMix_se(betahat_x, betahat_y, sx, sy, theta, pi0, sigma2)
}
\arguments{
\item{betahat_x}{GWAS effect estimates of the exposure, recommended to be in standardized scale. Vector of length \code{K}, where \code{K} is the number of instruments.}

\item{betahat_y}{GWAS effect estimates of the outcome, recommended to be in standardized scale. Vector of length \code{K}.}

\item{sx}{Standard error of \code{betahat_x}, recommended to be in standardized scale. Vector of length \code{K}.}

\item{sy}{Standard error of \code{betahat_y}, recommended to be in standardized scale. Vector of length \code{K}.}

\item{theta}{Estimate of causal effect.}

\item{pi0}{The probability mass of the null component corresponding to the estimated \code{theta}.}

\item{sigma2}{The variance of the non-null component corresponding to the estimated \code{theta}.}
}
\value{
The standard error of MRMix estimator.
}
\description{
This function calculates the standard error of the MRMix estimator using asymptotic theory.
}
\references{
Qi, Guanghao, and Nilanjan Chatterjee. "Mendelian randomization analysis using mixture models for robust and efficient estimation of causal effects." Nature Communications 10.1 (2019): 1941.
}
