% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize summary statistics for MRMix analysis}
\usage{
standardize(betahat_x, betahat_y, sx, sy, contbin_x, contbin_y, nx, ny,
  MAF)
}
\arguments{
\item{betahat_x}{GWAS effect estimates of the exposure. Vector of length \code{K}, where \code{K} is the number of instruments.}

\item{betahat_y}{GWAS effect estimates of the outcome. Vector of length \code{K}.}

\item{sx}{Standard error of \code{betahat_x}. Vector of length \code{K}.}

\item{sy}{Standard error of \code{betahat_y}. Vector of length \code{K}.}

\item{contbin_x}{Is the exposure a continuous or binary trait? Set to \code{contbin_x="continuous"} or \code{contbin_x="binary"}. Or set to \code{contbin_x="n"} if standardization is not needed.}

\item{contbin_y}{Is the outcome a continuous or binary trait? Set to \code{contbin_y="continuous"} or \code{contbin_y="binary"}. Or set to \code{contbin_y="n"} if standardization is not needed.}

\item{nx}{SNP-specific sample size of the study associated with the exposure. Vector of length \code{K} or a single number. A single number indicates all the SNPs have the same sample size. Set to \code{NULL} if trait is binary.}

\item{ny}{SNP-specific sample size of the study associated with the outcome. Vector of length \code{K} or a single number. A single number indicates all the SNPs have the same sample size. Set to \code{NULL} if trait is binary.}

\item{MAF}{Minor allele frequency. Vector of length \code{K}. Set to \code{NULL} if both traits are continuous.}
}
\value{
A list that contains
\item{betahat_x_std}{Standardized \code{betahat} for the exposure.}
\item{betahat_y_std}{Standardized \code{betahat} for the outcome.}
\item{sx_std}{Standard error of \code{betahat_x_std}.}
\item{sy_std}{Standard error of \code{betahat_y_std}.}
}
\description{
This function standardizes GWAS summary statistics (beta and standard error (SE)), i.e. 1) for continuous phenotype, transform them to the beta and SE when both the genotypes and phenotypes are standardized to have variance 1; 2) for binary phenotype, transform them to the beta and SE when the genotypes are standardized to have variance 1. This function applies to continuous phenotypes analyzed with linear regression and binary phenotypes analyzed with logistic regression. Do not use for other types of phenotypes or other analytic methods.
}
\details{
For continuous phenotypes analyzed with linear regression, data are standardized by \code{beta_standardized=beta/(se*sqrt(N)); se_standardized=1/sqrt(N)}; for binary phenotypes analyzed with logistic regression, data are standardized by \code{beta_standardized=beta*sqrt(2*MAF*(1-MAF)); se_standardized=se*sqrt(2*MAF*(1-MAF))}.
}
\references{
Qi, Guanghao, and Nilanjan Chatterjee. "Mendelian randomization analysis using mixture models for robust and efficient estimation of causal effects." Nature Communications 10.1 (2019): 1941.
}
